requirements.txt
Flask>=3.0.0
nbclient>=0.9.0
nbformat>=5.9.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
werkzeug>=3.0.0


